Changed format in Test_exp file
There's now an experimenter class. An Experimenter object contains the Neural network it uses for experiments, a data_manipulator object containing the data to be trained/tested on, and an integer specifying which method of training to be used (GA/ES/DE/Backprop). Currently 5x2 cross validation is the only train/test method, but may need more if data sets are too large.

Should be easy to fit other train methods onto this framewok